regression-test :: move more methods to ExtractionTest
