+++ src/test/java/emissary/test/core/junit5/ExtractionTest.java
@@ -5,9 +5,12 @@ import emissary.core.Family;
 import emissary.core.IBaseDataObject;
 import emissary.core.IBaseDataObjectXmlCodecs;
 import emissary.core.IBaseDataObjectXmlHelper;
+import emissary.core.channels.FileChannelFactory;
+import emissary.core.channels.SeekableByteChannelFactory;
 import emissary.kff.KffDataObjectHandler;
 import emissary.place.IServiceProviderPlace;
 import emissary.test.core.junit5.LogbackTester.SimplifiedLogEvent;
+import emissary.util.ByteUtil;
 import emissary.util.io.ResourceReader;
 import emissary.util.os.OSReleaseUtil;
 import emissary.util.xml.JDOMUtil;
@@ -17,6 +20,7 @@ import jakarta.annotation.Nullable;
 import jakarta.xml.bind.DatatypeConverter;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.jdom2.Attribute;
 import org.jdom2.DataConversionException;
@@ -32,10 +36,14 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URISyntaxException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Stream;
 
 import static emissary.core.IBaseDataObjectXmlCodecs.DEFAULT_ELEMENT_DECODERS;
@@ -73,6 +81,8 @@ public abstract class ExtractionTest extends UnitTest {
     private static final List<IBaseDataObject> NO_ATTACHMENTS = Collections.emptyList();
     private static final byte[] INCORRECT_VIEW_MESSAGE = "This is the incorrect view, the place should not have processed this view".getBytes();
 
+    private AnswerGenerator answerGenerator;
+
     /**
      * The list of actual logEvents generated by executing the place.
      */
@@ -107,6 +117,26 @@ public abstract class ExtractionTest extends UnitTest {
         }
     }
 
+    /**
+     * Override this or set the generateAnswers system property to true to generate XML for data files.
+     *
+     * @return defaults to false if no XML should be generated (i.e. normal case of executing tests) or true to generate
+     *         automatically
+     */
+    protected boolean generateAnswers() {
+        return Boolean.getBoolean("generateAnswers");
+    }
+
+    protected AnswerGenerator createAnswerGenerator() {
+        return null; // TODO :: create answer generator
+    }
+
+    protected AnswerGenerator getAnswerGenerator() {
+        if (answerGenerator == null) {
+            answerGenerator = createAnswerGenerator();
+        }
+        return answerGenerator;
+    }
 
     /**
      * This method returns the XML element decoders.
@@ -185,6 +215,34 @@ public abstract class ExtractionTest extends UnitTest {
     }
 
     /**
+     * Allow the initial IBDO to be overridden - for example, adding additional previous forms
+     * <p>
+     * This is used in the simple case to generate an IBDO from the file on disk and override the filename
+     *
+     * @param resource path to the dat file
+     * @return the initial IBDO
+     */
+    @Nullable
+    protected IBaseDataObject getInitialIbdo(final String resource) {
+        IBaseDataObject ibdo = new ClearDataBaseDataObject();
+        try {
+            final Path datFileUrl = Paths.get(new ResourceReader().getResource(resource).toURI());
+            final InitialFinalFormFormat datFile = new InitialFinalFormFormat(datFileUrl);
+            final SeekableByteChannelFactory sbcf = FileChannelFactory.create(datFile.getPath());
+            // Create a BDO for the data, and set the filename correctly
+            final IBaseDataObject initialIbdo = IBaseDataObjectXmlHelper.createStandardInitialIbdo(ibdo, sbcf, "Classification",
+                    datFile.getInitialForm(), kff);
+            initialIbdo.setChannelFactory(sbcf);
+            initialIbdo.setFilename(datFile.getOriginalFileName());
+
+            return initialIbdo;
+        } catch (final URISyntaxException e) {
+            fail("Couldn't get path for resource: " + resource, e);
+            return null;
+        }
+    }
+
+    /**
      * This method returns the logger name to capture log events from or null if log events are not to be checked.
      *
      * @return the logger name to capture log events from or null (the default) if log events are not to be checked.
@@ -248,6 +306,21 @@ public abstract class ExtractionTest extends UnitTest {
 
     protected void checkAnswersPreHook(Element answers, IBaseDataObject payload, IBaseDataObject attachment, String tname) {
         // Nothing to do here
+    }
+
+    /**
+     * Generates a SHA 256 hash of the provided bytes if they contain any non-printable characters
+     *
+     * @param bytes the bytes to evaluate
+     * @param alwaysHash overrides the non-printable check and always hashes the bytes.
+     * @return a value optionally containing the generated hash
+     */
+    protected Optional<String> hashBytesIfNonPrintable(byte[] bytes, final boolean alwaysHash) {
+        if (ArrayUtils.isNotEmpty(bytes) && (alwaysHash || ByteUtil.containsNonIndexableBytes(bytes))) {
+            return Optional.ofNullable(ByteUtil.sha256Bytes(bytes));
+        }
+
+        return Optional.empty();
     }
 
     /**
